Reorganized and fixed Get functions to pull all data regardless of Fragments.
Atrium Controllers keep Objects in a static position and does not change as Objects are deleted, which causes Fragmentations in the List of Objects. Previously, If any fragmentation existed in an Increment, the Get function would stop and return what it currently has, even though there could be more records in the Objects.